+++ src/transport/index.ts
@@ -1,3 +1,4 @@
 export * from './nostr-client-transport.js';
 export * from './nostr-server-transport.js';
+export * from './nostr-server/announcement-manager.js';
 export * from './base-nostr-transport.js';